🩹 Retrieve last patches from Feature/2020.q4.dev (#957)
* 🩹 2020.Q3.107 fixes (#455)

* 🩹 Closes #419

* 🎨 Closes #445

* 🩹 Closes #423

* 🩹 Closes #432

* 🩹 Closes #450

* 🩹 Closes #435

* 🩹 Fix user popup displayed for apps

* 🛠 Fix Sentry error-57

* 🩹 Fix incompatibilities matchAll and oldFileSystem

* 🇪🇸 Update es.js

Some minor language fixes

* Create user_story.md

* Update issue templates

* Update user-story.md

* Delete user_story.md

* 🩹 Patch 2 for 2020.q3.dev (#559)

* 🩹 2020.Q3.107 fixes (#455)

* 🩹 Closes #419

* 🎨 Closes #445

* 🩹 Closes #423

* 🩹 Closes #432

* 🩹 Closes #450

* 🩹 Closes #435

* 🩹 Fix user popup displayed for apps

* 🛠 Fix Sentry error-57

* 🩹 Fix incompatibilities matchAll and oldFileSystem

* Fix #548

* Fix #523

* Fix #553

* Fix #432

* Fix #558 - Quick implementation of Authorization header in Q3 version

* Close #552 Add image resizer

* Fix ternary priority issue

* Improve a bit reload hack

* Change black background to white

* Fix #480

* Fix all Drive inputs

* Fix #561

* Fix #479

* 🛠  Fix #479 and local disk thumbnail upload (#562)

* 🛠  Fix local disk thumbnail upload

* Fix connection timezon error

* Fix public drive access errors

* 🛠 Fix #479 fix #558

* 🛠 Fix issue related to local-storage saved inputs

* ⬆️ Update version

* 🐳 Update production workflow (#578)

* 🐳 Update production workflow

* Update saas-update-front.yml

* 🛠 Fix unwanted workspace redirect

* 🛠 Sentry patch on Q3 (#579)

* Update DrivePreviewCommand.php

* Update Notifications.php

* Update UsersConnections.php

* 🛠  Sentry patch

* 🔐  Fix security issue

* 🌟 Reduce notification reload impact on server

* 👟 Prepare production backend for new notification format

* 🐳 Fix compilation failure when building PHP Docker image (#611)

* 🛠  Realtime fixes (#944)

* Add segment in frontend

* Update notification backend endpoint

* Remove tag from frontend

* Do not read unfocused feeds

* Add logs for channel activity

* Do tests

* Fix channel activity not updating

* Fix unconfigured tab UI

* Remove TTL in channel cache for now

* Remove workspace creation rights for non admin / owners

* Add logs to websocket read error

* Add logs to secured connection and useWatcher to AppView

* Fix UI and reduce logs

* Fix infinite call for unknown user

* 🛠 Fixes and update version builder (#949)

* Update version builder

* Prepare changelog update

* Update migration script text

* Fix try catch not working

* Fix multi-bucket resumable issue

* Log more error for the error_service_not_found_or_not_allowed

* Fix

* Fix side view not updating

* 🛠 Multiple fixes including msg feed realtime issues

* Fix type thanks to @shivam9593

* Fix #807 and Snyk #947, #948

* Add code viewer

* Fix secured connection disconnections

* Fix package.lock

* 🛠 Multiple fixes including msg feed realtime issues (#954)

* 🤙 Fix calls and remove ephemeral messages (#955)

* 🛠 Fix "new version" detector (#956)

* Fix "new version" detector

* Fix Sentry issues

* Enable segment

* Save last channel openned in workspaces

* Fix creating jitsi call in event

* 🖐 Get changes from main

* 🔢 Set final version number

+++ twake/backend/core/src/Twake/Discussion/Services/MessageSystem.php
@@ -172,34 +172,38 @@ class MessageSystem
             $object["parent_message_id"] = "";
         }
 
-        $message_repo = $this->em->getRepository("Twake\Discussion:Message");
-        $message = $message_repo->findOneBy(Array("channel_id" => $object["channel_id"], "parent_message_id" => $object["parent_message_id"], "id" => $object["id"]));
+        if(!$object["ephemeral_id"]){
+            $message_repo = $this->em->getRepository("Twake\Discussion:Message");
+            $message = $message_repo->findOneBy(Array("channel_id" => $object["channel_id"], "parent_message_id" => $object["parent_message_id"], "id" => $object["id"]));
 
-        if (!$message) {
-            return false;
+            if (!$message) {
+                return false;
+            }
+
+            //TODO for allow_delete == "administrators" implement user access verification
+            if (!($this->hasAccess($message->getAsArray(), $current_user, $message) || $message->getAsArray()["hidden_data"]["allow_delete"] == "everyone" || $message->getAsArray()["hidden_data"]["allow_delete"] == "administrators")) {
+                return false;
+            }
+
+            if ($message->getResponsesCount() > 0) {
+                return false;
+            }
+
+            if ($message->getParentMessageId()) {
+                $parent_message = $message_repo->findOneBy(Array("channel_id" => $object["channel_id"], "parent_message_id" => "", "id" => $message->getParentMessageId()));
+                $parent_message->setResponsesCount($parent_message->getResponsesCount() - 1);
+                $this->em->persist($parent_message);
+            }
+
+            $array_before_delete = $message->getAsArray();
+
+            $this->em->remove($message);
+            $this->em->flush();
+
+            $this->deleteInBloc($message);
+        }else{
+            $array_before_delete = $object;
         }
-
-        //TODO for allow_delete == "administrators" implement user access verification
-        if (!($this->hasAccess($message->getAsArray(), $current_user, $message) || $message->getAsArray()["hidden_data"]["allow_delete"] == "everyone" || $message->getAsArray()["hidden_data"]["allow_delete"] == "administrators")) {
-            return false;
-        }
-
-        if ($message->getResponsesCount() > 0) {
-            return false;
-        }
-
-        if ($message->getParentMessageId()) {
-            $parent_message = $message_repo->findOneBy(Array("channel_id" => $object["channel_id"], "parent_message_id" => "", "id" => $message->getParentMessageId()));
-            $parent_message->setResponsesCount($parent_message->getResponsesCount() - 1);
-            $this->em->persist($parent_message);
-        }
-
-        $array_before_delete = $message->getAsArray();
-
-        $this->em->remove($message);
-        $this->em->flush();
-
-        $this->deleteInBloc($message);
 
         $event = Array(
             "client_id" => "system",
